service-python/authserver/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.1.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-suit==0.2.26
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
jedi==0.14.0
oauthlib==3.0.2
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-openid==3.1.0
requests==2.22.0
requests-oauthlib==1.2.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
